rest endpoins are created

+++ src/main/java/shippingservice/model/City.java
@@ -6,6 +6,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.util.Set;
 
 @NoArgsConstructor
 @AllArgsConstructor
@@ -19,8 +20,12 @@ public class City extends BaseModel{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private long id;
     private String city;
+
+//    @OneToMany
+//    @JoinColumn(name="city_id_fk")
+//    private Set<Area> areas;
 
     @ManyToOne(targetEntity=Region.class, cascade = CascadeType.MERGE)
     private Region region;
