+++ src/main/java/shippingservice/model/Region.java
@@ -7,6 +7,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.util.Set;
 
 @NoArgsConstructor
 @AllArgsConstructor
@@ -20,7 +21,13 @@ public class Region extends BaseModel{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private long id;
     private String region;
+
+//    @OneToMany
+//    @JoinColumn(name="region_id_fk")
+//    private Set<City> cities;
+
+
 
 }
